pyproject.toml
[project]
name = "ema-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.24.0",
    "instructor[google-genai]>=1.9.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.13.0",
    "sentence-transformers>=5.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "uv>=0.7.19",
]
